interface to 1kg import

interface to 1kg import

Usage

get_1kg(
  hl,
  retrieve_import_write = FALSE,
  path_1kg_zip = osn_1kg_path(),
  folder = tempdir(),
  cache = BiocFileCache::BiocFileCache(),
  simple = FALSE
)

Arguments

hl

hail object

retrieve_import_write

logical(1) if TRUE, use hl.utils.get_1kg to retrieve data, otherwise acquire a previously written zip file, either from a cache, or, if no file found in cache, from web, followed by caching

path_1kg_zip

character(1) path to zip of MatrixTable, defaults to `osn_1kg_path()`.

folder

character(1) destination of 1kg.mt as retrieved using hl.utils.get_1kg, import_vcf, write

cache

a BiocFileCache-type cache

simple

logical(1) if TRUE avoid basilisk to work with supplied hail module in `hl`, defaults to FALSE

Value

"hail.matrixtable.MatrixTable" instance

Note

overwrite is permitted in the import_vcf.write event

Examples

hl <- hail_init()
mt <- get_1kg(hl)
mt
#> <hail.matrixtable.MatrixTable object at 0x15c88ab20>
mt$rows()$select()$show(5L) # must use integer
annopath <- path_1kg_annotations()
tab <- hl$import_table(annopath, impute = TRUE)$key_by("Sample")
tab$describe()
tab$show(width = 100L)