update the reference genome for a hail instance

update the reference genome for a hail instance

Usage

rg_update(
  hc,
  init = "GRCh38",
  newjson = system.file("json/t2tAnVIL.json", package = "BiocHail")
)

Arguments

hc

hail context

init

character(1) valid name for a reference genome, defaults to "GRCh38"

newjson

character(1) path to a json spec of a reference genome [needs doc]

Value

a python list; the function is used for its side effect

Examples

hl <- hail_init()
rg_update(hl)
#> ReferenceGenome(name=GRCh38t2tb, contigs=['chr1', 'chr2', 'chr3', 'chr4', 'chr5', 'chr6', 'chr7', 'chr8', 'chr9', 'chr10', 'chr11', 'chr12', 'chr13', 'chr14', 'chr15', 'chr16', 'chr17', 'chr18', 'chr19', 'chr20', 'chr21', 'chr22', 'chrX', 'chrY', 'chrM'], lengths={'chr1': 248387497, 'chr2': 242696747, 'chr3': 201106605, 'chr4': 193575430, 'chr5': 182045437, 'chr6': 172126870, 'chr7': 160567423, 'chr8': 146259322, 'chr9': 150617274, 'chr10': 134758122, 'chr11': 135127772, 'chr12': 133324781, 'chr13': 114240146, 'chr14': 101219177, 'chr15': 100338308, 'chr16': 96330493, 'chr17': 84277185, 'chr18': 80542536, 'chr19': 61707359, 'chr20': 66210247, 'chr21': 45827691, 'chr22': 51353906, 'chrX': 154259625, 'chrY': 57227415, 'chrM': 16569}, x_contigs=['chrX'], y_contigs=['chrY'], mt_contigs=['chrM'], par=[('chrX', 10001, 2781480), ('chrX', 155701383, 156030896), ('chrY', 10001, 2781480), ('chrY', 56887903, 57217416)])